refactor: JwtProvider 패키지를 domain으로 변경
원하는 형식으로 JWT를 발급하고 검증하는 로직이 도메인 로직이라고 판단
