+++ src/main/java/com/github/teamreflog/reflogserver/auth/config/AuthenticationInterceptor.java
@@ -3,7 +3,7 @@ package com.github.teamreflog.reflogserver.auth.config;
 import static org.springframework.http.HttpMethod.OPTIONS;
 
 import com.github.teamreflog.reflogserver.auth.application.dto.AuthPrincipal;
-import com.github.teamreflog.reflogserver.auth.infrastructure.JwtProvider;
+import com.github.teamreflog.reflogserver.auth.domain.JwtProvider;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
