+++ src/test/java/com/github/teamreflog/reflogserver/auth/domain/JwtProviderTest.java
@@ -1,4 +1,4 @@
-package com.github.teamreflog.reflogserver.auth.infrastructure;
+package com.github.teamreflog.reflogserver.auth.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
